+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   * dumped `maven.home` and `maven.opts` configration - it's picked from the environment when used for War/Pom loading now
   * dumped `prepare` configuration; `create` for workspaces is no longer supported
   * dumped stage directory lock - stool now assumes that the project is used exclusively by the current user
+  * dumped Maven Embedded dependency
 * dumped `committed` configuration - start no longer checks for local modifications
 * removed `sleep` state and the corresponding `stop -sleep` flag, it was never used
 * removed `shared` switch - local maven repository is now always the user's Maven repository